implemented feature to add new students - with no save to file yet

+++ include/Student.h
@@ -16,7 +16,7 @@ public:
     Student(std::string name, int id);
     std::string getName();
     int getId();
-    int getGPA();
+    double getGPA();
     std::vector<int> getAllGrades();
 
     void updateName(std::string newName);
